rename acc_utils->acc_label, it only contains acc_label
