+++ bluepyopt/ephys/locations.py
@@ -23,7 +23,7 @@ Copyright (c) 2016-2020, EPFL/Blue Brain Project
 
 import itertools
 
-from bluepyopt.ephys.acc.acc_utils import ArbLabel
+from bluepyopt.ephys.acc.acc_label import ArbLabel
 from bluepyopt.ephys.base import BaseEPhys
 from bluepyopt.ephys.serializer import DictMixin
 from bluepyopt.ephys.parameterscalers import format_float
